+++ src/main/java/org/sunday/projectpop/controller/UserProfileController.java
@@ -1,4 +1,4 @@
-package org.sunday.projectpop.newnew.controller;
+package org.sunday.projectpop.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -6,10 +6,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import org.sunday.projectpop.newnew.dto.*;
-import org.sunday.projectpop.newnew.repository.UsersRepository;
-import org.sunday.projectpop.newnew.service.UserProfileService;
-import org.sunday.projectpop.newnew.service.UserTagCacheService;
+import org.sunday.projectpop.dto.ProfileResponse;
+import org.sunday.projectpop.dto.UserProfileRequestDTO;
+import org.sunday.projectpop.dto.UserProfileResponseDTO;
+import org.sunday.projectpop.dto.UserTagCacheDTO;
+import org.sunday.projectpop.repository.UsersRepository;
+import org.sunday.projectpop.service.UserProfileService;
+import org.sunday.projectpop.service.UserTagCacheService;
 
 import java.util.List;
 import java.util.UUID;
