+++ src/main/java/org/sunday/projectpop/dto/ProfileResponse.java
@@ -1,4 +1,4 @@
-package org.sunday.projectpop.newnew.dto;
+package org.sunday.projectpop.dto;
 
 public record ProfileResponse(
         UserProfileResponseDTO profile,
