+++ src/main/java/org/sunday/projectpop/dto/UserTagCacheDTO.java
@@ -1,4 +1,4 @@
-package org.sunday.projectpop.newnew.dto;
+package org.sunday.projectpop.dto;
 
 import java.util.List;
 import java.util.UUID;
