+++ src/main/java/org/sunday/projectpop/service/UserService.java
@@ -1,12 +1,12 @@
-package org.sunday.projectpop.newnew.service;
+package org.sunday.projectpop.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.sunday.projectpop.newnew.dto.JoinDTO;
-import org.sunday.projectpop.newnew.dto.LoginDTO;
-import org.sunday.projectpop.newnew.entity.Users;
-import org.sunday.projectpop.newnew.repository.UsersRepository;
+import org.sunday.projectpop.dto.JoinDTO;
+import org.sunday.projectpop.dto.LoginDTO;
+import org.sunday.projectpop.entity.Users;
+import org.sunday.projectpop.repository.UsersRepository;
 
 import java.util.UUID;
